// src/types/ProductType.ts
// Generated by https://quicktype.io

export interface ProductType {
    id_product: number;
    name: string;
    available_for_order: string;
    show_price: string;
    new_products: string;
    on_sale_products: string;
    quantity: number;
    minimal_quantity: string;
    allow_out_of_stock: string;
    float_price: number;
    price: string;
    discount_price: string;
    discount_percentage: string;
    images: Image[];
    cover_image: string;
    combinations: Combination[];
    options: Option[];
    description: string;
    description_short: string;
    reference: string;
    category_name: string;
    manufacturer_name: string;
    product_info: ProductInfo[];
    accessories: any[];
    customization_fields: CustomizationFields;
    pack_products: any[];
    seller_info: any[];
    product_attachments_array: any[];
    product_url: string;
    groups: Groups;
    breadcrumb: Breadcrumb[];
}

export interface Breadcrumb {
    title: string;
    url: string;
}

export interface Combination {
    id_product_attribute: number;
    quantity: number;
    price: string;
    float_price: number;
    minimal_quantity: number;
    combination_code: string;
}

export interface CustomizationFields {
    is_customizable: string;
    customizable_items: any[];
}

export interface Groups {
    "3": The3;
}

export interface The3 {
    group_name: string;
    name: string;
    group_type: string;
    default: number;
    attributes: { [key: string]: Attribute };
    attributes_quantity: { [key: string]: number };
}

export interface Attribute {
    name: string;
    html_color_code: string;
    texture: string;
    selected: boolean;
}

export interface Image {
    src: string;
}

export interface Option {
    id: number;
    title: string;
    is_color_option: number;
    items: Item[];
}

export interface Item {
    id: number;
    value: string;
    hex_value?: string
}

export interface ProductInfo {
    name: string;
    value: string;
}